for #660: split asserts.item
